pyproject.toml
[tool.poetry]
name = "cogs189"
version = "0.1.0"
description = ""
authors = ["Nick DeGroot <1966472+nickthegroot@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "~3.10"
numpy = "^1.22.1"
scipy = "^1.7.3"
matplotlib = "^3.5.1"
scikit-learn = "^1.0.2"
pyserial-asyncio = "^0.6"
reactivex = "^4.0.0-beta.3"
pandas = "^1.4.1"
seaborn = "^0.11.2"

[tool.poetry.dev-dependencies]
ipykernel = "^6.7.0"
black = {version = "^22.1.0", allow-prereleases = true}
torch-tb-profiler = "^0.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
